% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desire_high.R
\name{desire_high}
\alias{desire_high}
\title{High values are best}
\usage{
desire_high(x, cut1, cut2, min = 0, max = 1, scale = 1)
}
\arguments{
\item{x}{Vector of numeric values.}

\item{cut1, cut2}{Percentiles where the desirability function changes.}

\item{min, max}{Minimum (default = 0) and maximum (default = 1) desirability
scores.}

\item{scale}{Controls shape of the desirability function. Larger values
correspond to more steep and strict curves whereas smaller values correspond
to more gradual and inclusive curves.}
}
\value{
Returns a numeric vector of desirability scores.
}
\description{
This function maps a numeric variable to a [0, 1] scale where high values are
most desirable.
}
\details{
Values in the top n percent will have the highest possible
desirability and values in the botton n percent will have the lowest possible
desirability. Values between will be transformed to intermediate values
according to the function.
}

