% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desire_plot.R
\name{desire_plot}
\alias{desire_plot}
\title{Plot desirability scores}
\usage{
desire_plot(x, plot_type = plot.type)
}
\arguments{
\item{x}{A data matrix where the first column is overall desirability and
additional columns are desirability scores from individual tests.}

\item{plot_type}{Type of plot to create (overall, top, type, study, or all).}
}
\value{
Returns a geom_point plot generated by ggplot.
}
\description{
This function plots genes with the top overall desirability scores and,
optionally, their desirability scores from all integrated data.
}
\details{
By plotting genes with the top overall desirability scores as well
as their individual desirability scores from integrated data, the user can
visualize not only the most desirable candidates, but also where the
desirability signal is coming from within their data.
}
