% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riemann.R
\name{riemann}
\alias{riemann}
\title{Approximate an integrable function by Riemann sum}
\usage{
riemann(f, range, n)
}
\arguments{
\item{f}{a function to be integrated}

\item{range}{integration range}

\item{n}{number of intervals}
}
\value{
integration result
}
\description{
Approximate an integrable function by Riemann sum
}
\examples{
# A function to be integrated:
f <- function(x) {return(sqrt(x))}
res <- riemann(f, c(0,1), 10)

}
