% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpsons.R
\name{simpsons}
\alias{simpsons}
\title{Simpsons method}
\usage{
simpsons(f, range, n)
}
\arguments{
\item{f}{a function to be integrated}

\item{range}{integration range}

\item{n}{number of intervals}
}
\value{
integration result
}
\description{
Simpsons method
}
\examples{
# A function to be integrated:
f <- function(x) {return(sqrt(x))}
res <- simpsons(f, c(0,1), 10)

}
