% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trapezoid.R
\name{trapezoid}
\alias{trapezoid}
\title{Trapezoid rule}
\usage{
trapezoid(f, range, n)
}
\arguments{
\item{f}{a function to be integrated}

\item{range}{integration range}

\item{n}{number of intervals}
}
\value{
integration result
}
\description{
Trapezoid rule
}
\examples{
# A function to be integrated:
f <- function(x) {return(sqrt(x))}
res <- trapezoid(f, c(0,1), 10)

}
